Added a new FileDownloader method, modifyMimickedCookieDomain.
This allows the caller to find a particular cookie that is being
mimicked and update the cookie to use a different domain. This
method was added to workaround an apparent InternetExporerDriver
defect. In my testing, the IE WebDriver ignored the domain for
cookies, and passed them as null to the FileDownloader; this
caused tests to fail for having invalid cookies.