Refatored code to check and verify the domain for a particular cookie.
No need to have the code loop through the mimic cookies twice; instead,
check for and update the domain of a cookie the first (and only) time
through. If the user does not call verifyCookieDomainWhenDownloading or
verifyCookieDomainWhenCheckingStatus, then the domain checking/setting
is skipped. This code fixes the isues I had with IE webdriver "loosing"
the domain portion of cookies.

Limitation: Only works for one cookie per download/urlChecker object.
Other: The code is basically duplicated between the two classes. I
did this to keep things in alignment with the original codebase. However,
this could be refactored to pull the reusable bits and have them only
in one place, not two.